pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.4,<3.13.0"
dependencies = [
    "chromadb>=1.0.4",
    "dashscope>=1.23.1",
    "faiss-cpu>=1.10.0",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "markdown>=3.8",
    "openai>=1.73.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pdfplumber>=0.11.6",
    "pip>=25.0.1",
    "pypdf>=5.4.0",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "sentence-transformers>=4.0.2",
    "setuptools>=78.1.0",
    "spacy[apple]>=3.8.5",
    "unstructured>=0.17.2",
    "wheel>=0.45.1",
]
